requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
praw==7.7.1
nltk==3.8.1
python-dotenv==1.0.0
numpy>=1.26.0
pandas>=2.1.0
regex>=2021.8.3
scikit-learn>=1.3.0
gensim>=4.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0
vaderSentiment>=3.3.2 